A whole loada stuff
U.I text is now white to aid visibility
Apothecary combat now works correctly
Player has a mesh in the form of a T-posing ethan
Door to the boss rom is now correctly locked
There are now blockers in the junction room that have to be whacked with either the sword or shield in order to destroy them. These also prevent players from jumping past via teleport points. - Note: Sword one doesn't quite work atm
Boss fight begun to be implemented. There's a stream of light particles falling from the celling that the player needs to deflect into the boss' plinth in order to bring them down in order to fight(Doesn't quite detect the collisions atm). The boss fight is as standard, but with more health. Upon defeating it, the last few points will activate and the door to the treasure room will open